[add] SQLite integration with diesel & rocket_contrib

+++ src/prime/utils.rs
@@ -1,14 +1,19 @@
 use rocket::response::content::Json;
 
+// pub struct DBPrime {
+//     pub id: i32,
+//     pub value: i64,
+// }
+
 pub struct Prime {
     pub value: u64,
-    pub position: usize,
+    pub position: u64,
     pub new: bool,
     pub duration: f64,
 }
 
 impl Prime {
-    pub fn new(value: u64, position: usize, duration: f64, new: bool) -> Prime {
+    pub fn new(value: u64, position: u64, duration: f64, new: bool) -> Prime {
         Prime {
             value: value,
             position: position,
